Now compiling with Java 8 time library and JodaTime removed. Non of the tests are expected to run green yet.

+++ meteo-extras/src/main/java/no/api/meteo/services/ScoreForecast.java
@@ -24,8 +24,8 @@ class ScoreForecast {
 
     private final PeriodForecast periodForecast;
 
-    private final int pointTightScore;
+    private final long pointTightScore;
 
-    private final int pointWideScore;
+    private final long pointWideScore;
 
 }
